Make mbedtls_timing_hardclock use unsigned integer
When calling mbedtls_timing_hardclock on a platform
that does not have a platform-specific implementation,
it multiplies the difference in seconds since start
with 1_000_000 to get the microseconds. After 36 minutes,
this will cause a signed integer overflow.
A signed integer overflow will cause a trap when the
GCC -ftrapv hardening flag is set.

This fix converts the constant to unsigned, so that the
multiplication becomes unsigned and won't trap on overflow.
Note that an overflow is allowed by-design.

Closes #3066